removed cadent rep distributor from repository

+++ src/RepTokens.sol
@@ -10,8 +10,7 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
     bytes32 public constant MINTER_ROLE = keccak256("MINTER_ROLE");
     bytes32 public constant DISTRIBUTOR_ROLE = keccak256("DISTRIBUTOR_ROLE");
     bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
-    bytes32 public constant TOKEN_MIGRATOR_ROLE =
-        keccak256("TOKEN_MIGRATOR_ROLE");
+    bytes32 public constant TOKEN_MIGRATOR_ROLE = keccak256("TOKEN_MIGRATOR_ROLE");
 
     uint256 public maxMintAmountPerTx;
     mapping(uint256 => address[]) ownersOfTokenTypes;
@@ -20,23 +19,13 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
     event Mint(address minter, address to, uint256 amount);
     event DestinationWalletSet(address coreAddress, address destination);
     event Distributed(address from, address to, uint256 amount);
-    event OwnershipOfTokensMigrated(
-        address from,
-        address to,
-        uint256 lifetimeBalance,
-        uint256 redeemableBalance
-    );
+    event OwnershipOfTokensMigrated(address from, address to, uint256 lifetimeBalance, uint256 redeemableBalance);
     event BurnedRedeemable(address from, address to, uint256 amount);
 
     //id 0 = lifetime token
     //id 1 = transferable token
-    constructor(
-        address[] memory admins,
-        uint256 _maxMintAmountPerTx
-    )
-        ERC1155(
-            "ipfs://bafybeiaz55w6kf7ar2g5vzikfbft2qoexknstfouu524l7q3mliutns2u4/{id}"
-        )
+    constructor(address[] memory admins, uint256 _maxMintAmountPerTx)
+        ERC1155("ipfs://bafybeiaz55w6kf7ar2g5vzikfbft2qoexknstfouu524l7q3mliutns2u4/{id}")
     {
         for (uint256 i = 0; i < admins.length; i++) {
             _setupRole(DEFAULT_ADMIN_ROLE, admins[i]);
@@ -45,32 +34,18 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         maxMintAmountPerTx = _maxMintAmountPerTx;
     }
 
-    function uri(
-        uint256 _tokenid
-    ) public pure override returns (string memory) {
-        return
-            string(
-                abi.encodePacked(
-                    "ipfs://bafybeiaz55w6kf7ar2g5vzikfbft2qoexknstfouu524l7q3mliutns2u4/",
-                    Strings.toString(_tokenid)
-                )
-            );
+    function uri(uint256 _tokenid) public pure override returns (string memory) {
+        return string(
+            abi.encodePacked(
+                "ipfs://bafybeiaz55w6kf7ar2g5vzikfbft2qoexknstfouu524l7q3mliutns2u4/", Strings.toString(_tokenid)
+            )
+        );
     }
 
-    function mint(
-        address to,
-        uint256 amount,
-        bytes memory data
-    ) public onlyRole(MINTER_ROLE) whenNotPaused {
-        require(
-            amount <= maxMintAmountPerTx,
-            "Cannot mint that many tokens in a single transaction!"
-        );
+    function mint(address to, uint256 amount, bytes memory data) public onlyRole(MINTER_ROLE) whenNotPaused {
+        require(amount <= maxMintAmountPerTx, "Cannot mint that many tokens in a single transaction!");
 
-        require(
-            hasRole(DISTRIBUTOR_ROLE, to),
-            "Minter can only mint tokens to distributors!"
-        );
+        require(hasRole(DISTRIBUTOR_ROLE, to), "Minter can only mint tokens to distributors!");
 
         //mints an amount of lifetime tokens to an address.
         super._mint(to, 0, amount, data);
@@ -80,19 +55,17 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         emit Mint(_msgSender(), to, amount);
     }
 
-    function mintBatch(
-        address[] memory to,
-        uint256[] memory amount,
-        bytes memory data
-    ) public onlyRole(MINTER_ROLE) whenNotPaused {
+    function mintBatch(address[] memory to, uint256[] memory amount, bytes memory data)
+        public
+        onlyRole(MINTER_ROLE)
+        whenNotPaused
+    {
         for (uint256 i = 0; i < to.length; i++) {
             mint(to[i], amount[i], data);
         }
     }
 
-    function setMaxMintAmount(
-        uint256 value
-    ) public onlyRole(DEFAULT_ADMIN_ROLE) {
+    function setMaxMintAmount(uint256 value) public onlyRole(DEFAULT_ADMIN_ROLE) {
         maxMintAmountPerTx = value;
     }
 
@@ -100,22 +73,18 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         _setDestinationWallet(_msgSender(), destination);
     }
 
-    function _setDestinationWallet(
-        address coreAddress,
-        address destination
-    ) internal {
+    function _setDestinationWallet(address coreAddress, address destination) internal {
         destinationWallets[coreAddress] = destination;
         emit DestinationWalletSet(coreAddress, destination);
     }
 
     //from : distributor
     //to : address
-    function distribute(
-        address from,
-        address to,
-        uint256 amount,
-        bytes memory data
-    ) public onlyRole(DISTRIBUTOR_ROLE) whenNotPaused {
+    function distribute(address from, address to, uint256 amount, bytes memory data)
+        public
+        onlyRole(DISTRIBUTOR_ROLE)
+        whenNotPaused
+    {
         if (destinationWallets[to] == address(0)) {
             _setDestinationWallet(to, to);
         }
@@ -125,12 +94,11 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         emit Distributed(from, destinationWallets[to], amount);
     }
 
-    function distributeBatch(
-        address from,
-        address[] memory to,
-        uint256[] memory amount,
-        bytes memory data
-    ) public onlyRole(DISTRIBUTOR_ROLE) whenNotPaused {
+    function distributeBatch(address from, address[] memory to, uint256[] memory amount, bytes memory data)
+        public
+        onlyRole(DISTRIBUTOR_ROLE)
+        whenNotPaused
+    {
         for (uint256 i = 0; i < to.length; i++) {
             distribute(from, to[i], amount[i], data);
         }
@@ -138,13 +106,10 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
 
     //from : address
     //to : burner
-    function safeTransferFrom(
-        address from,
-        address to,
-        uint256 id,
-        uint256 amount,
-        bytes memory data
-    ) public override {
+    function safeTransferFrom(address from, address to, uint256 id, uint256 amount, bytes memory data)
+        public
+        override
+    {
         require(id == 1, "Can only send a redeemable token!");
 
         require(
@@ -154,10 +119,7 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
 
         require(!hasRole(BURNER_ROLE, from), "Burners cannot send tokens!");
 
-        require(
-            hasRole(BURNER_ROLE, to),
-            "Can only send Redeemable Tokens to burners!"
-        );
+        require(hasRole(BURNER_ROLE, to), "Can only send Redeemable Tokens to burners!");
 
         super.safeTransferFrom(from, to, id, amount, data);
         emit BurnedRedeemable(from, to, amount);
@@ -172,7 +134,7 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         bytes memory data
     ) internal override {
         //loop through transferred token IDs
-        for (uint i = 0; i < ids.length; i++) {
+        for (uint256 i = 0; i < ids.length; i++) {
             //if the tokenID balance of the receiving address is greater than zero after the transfer, then check to see if the receiving
             //address needs to be added as an owner to the tokenID
             if (balanceOf(to, ids[i]) > 0) {
@@ -194,30 +156,19 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
 
     //this needs to be called beforehand by address that wants to transfer its lifetime tokens:
     //setApprovalForAll(TOKEN_MIGRATOR_ROLE, true)
-    function migrateOwnershipOfTokens(
-        address from,
-        address to
-    ) public onlyRole(TOKEN_MIGRATOR_ROLE) {
+    function migrateOwnershipOfTokens(address from, address to) public onlyRole(TOKEN_MIGRATOR_ROLE) {
         uint256 lifetimeBalance = balanceOf(from, 0);
         uint256 redeemableBalance = balanceOf(from, 1);
 
         super.safeTransferFrom(from, to, 0, lifetimeBalance, "");
         super.safeTransferFrom(from, to, 1, redeemableBalance, "");
-        emit OwnershipOfTokensMigrated(
-            from,
-            to,
-            lifetimeBalance,
-            redeemableBalance
-        );
+        emit OwnershipOfTokensMigrated(from, to, lifetimeBalance, redeemableBalance);
     }
 
     //@addrToCheck: Address to check during _afterTokenTransfer if it is already registered
     //as an owner of @tokenID.
     //@tokenID: the ID of the token selected.
-    function addAddressAsOwnerOfTokenIDIfNotAlreadyPresent(
-        address addrToCheck,
-        uint256 tokenID
-    ) internal {
+    function addAddressAsOwnerOfTokenIDIfNotAlreadyPresent(address addrToCheck, uint256 tokenID) internal {
         //get all owners of a given tokenID.
         address[] storage owners = ownersOfTokenTypes[tokenID];
 
@@ -242,35 +193,28 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         }
     }
 
-    function removeAddressAsOwnerOfTokenID(
-        address addrToCheck,
-        uint256 id
-    ) internal {
+    function removeAddressAsOwnerOfTokenID(address addrToCheck, uint256 id) internal {
         address[] storage owners = ownersOfTokenTypes[id];
 
         uint256 index;
-        for (uint i = 0; i < owners.length; i++) {
+        for (uint256 i = 0; i < owners.length; i++) {
             if (owners[i] == addrToCheck) {
                 index = i;
                 break;
             }
         }
 
-        for (uint i = index; i < owners.length - 1; i++) {
+        for (uint256 i = index; i < owners.length - 1; i++) {
             owners[i] = owners[i + 1];
         }
         owners.pop();
     }
 
-    function getOwnersOfTokenID(
-        uint256 tokenID
-    ) public view returns (address[] memory) {
+    function getOwnersOfTokenID(uint256 tokenID) public view returns (address[] memory) {
         return ownersOfTokenTypes[tokenID];
     }
 
-    function getOwnersOfTokenIDLength(
-        uint256 tokenID
-    ) public view returns (uint256) {
+    function getOwnersOfTokenIDLength(uint256 tokenID) public view returns (uint256) {
         return ownersOfTokenTypes[tokenID].length;
     }
 
@@ -282,9 +226,13 @@ contract RepTokens is AccessControl, Ownable, ERC1155, Pausable {
         }
     }
 
-    function supportsInterface(
-        bytes4 interfaceId
-    ) public view virtual override(ERC1155, AccessControl) returns (bool) {
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC1155, AccessControl)
+        returns (bool)
+    {
         return super.supportsInterface(interfaceId);
     }
 }
